requirements.txt
transformers
diffusers
accelerate
huggingface-hub
ninja
xatlas
imageio
imageio-ffmpeg
matplotlib
pyrallis
loguru
tqdm
einops
opencv-python
hydra-core
pillow
bson
trimesh
numpy
flask
flask_cors 
openai
boto3
objaverse